ajouter with relation

+++ backend/database/migrations/2026_04_03_112337_create_reservations_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->date('date_enter');
             $table->date('date_sortie');
             $table->integer('chamber');
-            $table->float('prix');
+            $table->numeric('prix');
             $table->integer('user_id');
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->integer('chamber_id');
